Renamed ExtensionMgr to ExtensionManager
