+++ src/brackets.js
@@ -105,7 +105,7 @@ define(function (require, exports, module) {
     require("search/FindInFiles");
     require("search/FindReplace");
     require("extensibility/InstallExtensionDialog");
-    require("extensibility/ExtensionMgrDialog");
+    require("extensibility/ExtensionManagerDialog");
     
     PerfUtils.addMeasurement("brackets module dependencies resolved");
