+++ src/extensibility/extension-manager-dialog.html
@@ -1,4 +1,4 @@
-<div class="extension-mgr-dialog modal">
+<div class="extension-manager-dialog modal">
     <div class="modal-header">
         <h1 class="dialog-title">{{EXTENSION_MANAGER_TITLE}}</h1>
     </div>
