+++ src/extensibility/extension-manager-view.html
