+++ src/styles/brackets_patterns_override.less
@@ -464,7 +464,7 @@
 }
 
 /* Extension Manager dialog */
-.extension-mgr-dialog .modal-body {
+.extension-manager-dialog .modal-body {
     width: 700px;
     height: 500px;
     text-align: center;
