+++ test/UnitTestSuite.js
@@ -38,7 +38,7 @@ define(function (require, exports, module) {
     require("spec/EditorOptionHandlers-test");
     require("spec/EditorManager-test");
     require("spec/ExtensionInstallation-test");
-    require("spec/ExtensionMgr-test");
+    require("spec/ExtensionManager-test");
     require("spec/ExtensionUtils-test");
     require("spec/FileIndexManager-test");
     require("spec/FindReplace-test");
