+++ test/spec/ExtensionManager-test-files/mockRegistry.json
